Fahr to celsius table over 20, using celsius = 5*(fahr-32)/9, p.25

+++ main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
-//Fahr to celsius table over 20, using celsius = 5*(fahr-32)/9, p.25
+//Fahr to celsius table over 20, using celsius = 5*(fahr-32)/9, p.25 2nd ver with float
 int main() {
-    int fahr, celsius;
+    float fahr, celsius;
     int lower, upper, step;
     
     lower = 0;
@@ -11,8 +11,8 @@ int main() {
     printf("Fahrenheit:\tCelsius:\n");
     fahr = lower;
     while (fahr <= 300) {
-        celsius = 5 * (fahr - 32) /9;
-        printf("%d\t\t%d\n", fahr, celsius);
+        celsius = 5.0 * (fahr - 32.0) /9.0;
+        printf("%.1f\t\t%.1f\n", fahr, celsius);
         fahr = fahr + step;
     }
     return 0;
